Pause and resume game functions
Added in a PlayerStateHandler class to allow the player to pause and resume the game. The code will, when the player presses the pause button, efficiently freeze time, unlock the player’s cursor and swap out the heads up display canvas for the pause menu one. Pressing the pause button again will efficiently undo these changes. This functionality will make it much easier to playtest the game in Edit mode without moving the mouse cursor off-screen and risking the player accidentally clicking something else while trying to play. Pausing and resuming also have UnityEvents attached, allowing extra functionality to easily be added, such as disabling the character controller.
This code was technically added in a previous commit, but I am including the information in this log since these files were technically updated due to being moved to a different folder.