Entity system
Did a small amount of work creating an Entity class, as well as some derived variants. These classes are used as a way to easily reference objects in the scene that meaningfully contribute to gameplay by including relevant statistics, with inherited versions used to determine specific kinds.
For example, the Player class contains references to important player-related scripts, as a central handler to allow them to easily communicate with and reference each other. Then any function in the game that needs to look for a player can easily search for said class and gain access to all its important variables.