Start of weapon and firing mode system
Created an abstract WeaponMode class. This mode will have inherited variants for different kinds of attacks, e.g. ranged, melee, or otherwise. These classes will contain important attack code and stats, and can be slotted in to give weapons a wide variety of different attacks.
Created a Weapon MonoBehaviour class for equippable weapons. This class does not contain any attack code, but includes code for drawing and holstering, as well as an array of WeaponModes to switch between.