Health system and hitboxes
Created a Health MonoBehaviour for storing standard health stats, such as max and current capacities, and code for dealing damage.
I also implemented a Hitbox MonoBehaviour. This class attaches to a collider and will reference a Health class. When a damaging attack hits a collider, it will check for one of these classes, and deal damage to its corresponding Health class. The Health class, during its Awake() function, will automatically detect its DamageHitboxes using GetComponentsInChildren, and assign itself to each one. This separation between hitbox and health data is necessary for entities with complex shapes that require multiple hitboxes, and for giving different parts of an entity’s body different reactions to attacks. For example, the hitbox on the head of a humanoid enemy will have a critical variable set to true, so attacks will multiply their damage accordingly.