PointDamage and placeholder bullet prefab
I created a PointDamage class for basic projectiles that deal damage to a single point on impact. This is a MonoBehaviour that attaches to a Projectile object, and is triggered by a UnityEvent. It contains a base damage value, a multiplier for critical hits, and two UnityEvents – one for when the projectile successfully damages, and one for if the object is unharmed (based on factors such as if the object has a proper damage hitbox, or if they’re an ally and can’t be hurt by your shots).
These functions are separate for increased modularity, in case a projectile needs to exist that does not deal damage, or does so via some other means. For example, a rocket or grenade would instead have an area-of-effect damage function that involves checking in a radius.
I also assembled a placeholder bullet prefab using these classes, to test the functionality.