KillBounds class for deleting clutter from levels
I created a simple MonoBehaviour class named KillBounds. A single one of these will exist in the game level, and will run a sweep every few seconds. This sweep will detect all entities inside the level, and check if they are out of bounds (specified by a Bounds struct). If the entity is outside the level bounds, it will be killed or otherwise deleted, to prevent useless objects from cluttering up the level while not contributing to gameplay, such as stray projectiles that sail over the horizon.