Resource struct
I made a custom struct for managing resources. It contains values for the maximum and current capacities, as well as a third value for a critical threshold, to determine certain behaviours such as HUD indicators and enemy AI. I made it so I didn’t have to deal with the trouble of writing three separate variables each time I wanted some kind of value that depleted and refilled.