Noise-based variable accuracy and fixing drawing functions
Updated the WeaponHandler class to include a Vector3 property for the player’s aim direction. Rather than random angles, the function combines two Perlin noise values (determined using Time.time) with the player’s accuracy sway angle to produce a Vector3 that slowly shifts around within the player’s accuracy constraints. The version I had in my previous project would randomly decide angles. This new version should have three benefits – one, it’ll look cooler – two, it’ll be more realistic – and three, it should theoretically encourage a different playstyle.
My theory: when shots are spraying in a completely random distribution, with fast-firing automatic weapons it can be easier to spray and pray because the shots are covering a wider area. When the aim direction is not theoretically anywhere but slowly shifting from one point to another, it will be less viable to hold down the fire button and carpet an area in shots, because the shots will take time to actually reach that point. Therefore, players will be incentivised to actually aim their shots.
I decided to make this change because in my previous version of the game, the assault rifle was overpowered. Despite being designed for longer-range shooting, its automatic fire and spread made it easy for players to use in close-quarters skirmishes, somewhat defeating the point of the other weapons. Ideally, crimping the weapon’s accuracy this way should make it harder to use in close quarters, but easier to calculate precise shots (at least when I add some extra code to ensure the reticle accurately represents where the swayed direction is pointing). For weapons that should actually be used in close quarters, there are separate calculations for spread, which does use the old random distribution system (although I want to figure out a better method that ensures more even distribution than that).

I also changed the verification function for the WeaponHandler (the one that determines what weapons are currently equipped on the player and updates everything accordingly) to automatically run the Draw function on the ‘current’ weapon, since there is presently no input to do it deliberately.