Ammunition Inventory
Made an AmmunitionInventory MonoBehaviour to manage quantities and capacities for the player’s ammunition reserves. This code is very similar to the version in my old project, but with some minor changes to account for the Resource struct, which was a class in my previous project.
Firstly, there is an AmmunitionType enum used to easily determine the appropriate type, for purposes such as ammunition consumption and pickups that restore specific types. To ensure the right amount of fields are present for the amount of types, an array exists that is automatically filled in during the OnValidate function. The code will get a list of names based off the enum’s values, create a new Resource struct array, and check them against the current entries. If an entry is not present, it will create a new one. If an entry already exists and has values filled in, it will determine said info via the Resource’s name value and copy the values over to the new array. The new array will then overwrite the old one, deleting any redundant entries.
This system is a foolproof way to update the array, adding new types and removing old ones without deleting information for existing types. For consuming and restoring ammunition, the Spend and Collect functions are used, which use the AmmunitionType enum as parameters to select the appropriate array. The only change I may make is to change the enum to a ScriptableObject, allowing me to add new parameters for each type, such as a description and HUD icon.