Updating HUD resource meter (untested)
I created, but have not tested, a prefab for a resource meter to appear on a heads-up display. The meter consists of text, a bar for the amount, and a secondary bar that will smoothly move over time (using an IEnumerator) to match the main bar, to show large chunks of damage being dealt.