Updated gun fire controller to consume ammo properly
I added a special bool to properly check if the player’s gun can be fired, which is referenced in the fire controller’s IEnumerator whenever it is time to shoot. Currently it just contains code to check if the magazine and ammunition reserve have ammo present, but due to its modular nature other checks can be easily added. I also added code to the firing sequence to deplete the appropriate amount and type of ammunition. The ammunition code works well, but I have not properly tested the magazine code yet.