Added simple weapon animation functionality
Added code into the LateUpdate() loop of my Weapon class, to lerp the weapon’s model between certain points for simple animation, for purposes such as drawing, holstering, firing, reloading and/or aiming down sights. The positions are determined by a SimpleWeaponAnimation MonoBehaviour class, containing information for an old position, a new position, a time value and an animation curve for controlling the movement’s smoothness. The system needs polishing, but it works well enough that I have faith in the general idea. And even if it doesn’t work that well, its main purpose is to be playable for a prototype.