Added UnityEvents in WeaponHandler to trigger functions that apply to all weapons, and modified swaying aim direction to be easier to use
Added UnityEvents to the WeaponHandler to fire when the player draws, holsters, fires or switches the mode of any weapon. This was implemented for purposes such as updating HUD elements that would apply to any weapon.
Modified aim direction value to automatically account for the equipped weapon’s sway value, without needing to be included as a separate parameter.