Player prefab changes – restructured player transform hierarchy for ADS sway
I altered the hierarchy to mount the player camera and arms on top of the transform used for aiming, like a ball and socket joint. This will exist for the purpose of allowing aim sway to affect the player's camera so it lines up with the aim direction when using ADS functionality, for more precise aiming. I also tried making some code in the GunOptics class to implement this sway, rotating the player’s camera and arms to match the WeaponHandler’s aim direction value, but I have not properly implemented or tested it yet.