Weapon switching implemented and running smoothly
I have now implemented functionality to switch weapons, which seems to work perfectly. The system uses an IEnumerator that triggers the Holster() IEnumerator for the current weapon, waits until said weapon’s isSwitching bool is set back to false, then changes the firing mode index and triggers the Draw() IEnumerator for whatever is now the current weapon. For compatibility, the current index is clamped to always fit inside the array, and if the array does not have any entries, the function ends prematurely since there is nothing to switch to. If the index already equals the current weapon, and it is already active, the function also ends prematurely. It checks the active state in case the player has no weapons drawn and the one they are technically switched to is inactive.
For switching controls, I initially implemented a simple scroll function that altered the firing mode index by single increments to cycle through each available weapon. I then realised that I wanted that input to be used for firing modes, so I also created a simply hotkey system that checks if the player has pressed a number key on their keyboard and returns an appropriate integer. This simple system allows easy number key-based hotbar switching, a common feature in many shooters. I intend on adding other control methods later down the line, primarily a radial selector to make my game more compatible with controllers.
I also removed a dedicated WaitForSeconds reference in the Weapon class, because I felt the performance gain of not making a new one every time the player switches weapons wasn’t worth the added complexity and messier code.