Button input structs
I made a ButtonInput struct containing a string referencing a button in the Input manager. This struct then has three bools for being pressed, held and released, that reference the status of a particular button. This acts as an easy way to get all three states of a particular button without typing one of the three regular functions, especially when a single button’s status is recorded lots of times for a particular purpose.
This also allows me to have a more modular system for managing attack inputs. I included three in my WeaponHandler class, for primary (left mouse button/right trigger), secondary (right mouse button/left trigger), and tertiary (R/X) weapon functions, which I will be able to easily change the references to if I decide to overhaul the input system (a possibility considering the awkward problems with Unity’s default input system).
I forgot to include this info when pushing the commit with the WeaponHandler class, so I’m putting it here since this is one of the classes that contains a reference to it and it’s the only change that’s been made. I’m slightly embarrassed about it since it’s a rookie mistake, but I’m typing it here so the change is still documented.