Tweaked heads-up display to fix some bugs and require less setup
I changed the weapon HUD code to be run continuously in LateUpdate(), as the performance optimisation of making it run only when needed wasn’t worth the complexity of adding triggers for every single scenario that could cause it to change.
I also added a reference to the player’s health class in the Player class, allowing the heads-up display to automatically add listeners to update the health meter, without the developer having to add them manually.
Finally, I added an extra line of code to alter the ‘max’ value of the ammunition reserve meter, to not include ammunition that was supposed to be in the magazine. Without this change, the meter would appear less than full even though when taking the current magazine into account the numerical amount would be correct.