Moved reload controls into GunMagazine for increased modularity, and cleaned up logic code to be easier to read with less redundant lines.
The reloading controls and checking for the magazine being empty are now included in a function named InputLoop, contained inside the GunMagazine class and ran from the RangedAttack class. The increased modularity should make it easier to implement magazine and reloading functionality into other contexts.
I also reorganised several functions and variables in the RangedAttack class.