Simple ADS functionality
My original plan was to use IEnumerators to increase modularity, but I found that for the ADS code it was far better to just have a continuous update loop for managing the functionality (albeit all self-contained inside the class for improved modularity). The loop starts with a simple check of the player’s input, then if the timer value is not equal to the desired value for the state (1 if using ADS, 0 if not, determined by an easy to use property), lerping functions will be run to shift the player’s FOV between its default value and a new value divided to achieve the proper magnification. The preliminary timer value check is present to ensure these functions are not unnecessarily run every frame.
What is run every frame, however, is code for showing swaying aim while aiming down sights. The player’s aim direction is already determined using the WeaponHandler’s AimDirection() function, which produces a swaying Vector3 direction for a more accurate representation of the player’s swaying weapon. While aiming down the sights, the player’s camera will also rotate to face in this direction, allowing the player to accurately view and compensate for the sway. This will allow skilled players to compensate for the aim sway and improve their accuracy.
I also implemented a better check to prevent the WeaponHandler from switching weapons while the player is currently performing an action. I had to implement this to ensure that the player cannot switch weapons until all weapon functions have finished and returned to their neutral states, e.g. so the player cannot switch while aiming down the sights of a weapon and be stuck with zoomed in view and wonky sensitivity on another weapon that does not have ADS functionality.
There are still other features I need to implement, such as altering movement speed and camera sensitivity, moving the player’s weapon model and implementing nicer-looking scope overlays. But I am quite pleased with the system I have currently, as it works fluidly with no bugs that I have presently noticed.