Imported gunshot debris effect from my previous project.
Now the guns make pleasing impact particles when hitting surfaces!
The prefab had a missing script intended to destroy the object after a short period of time, since the effect is instantiated when a bullet hits a surface. I removed this and added my new LimitedLifetimeObject class to achieve the same effect.