Fully implemented functions for spawning objects upon impact
I implemented a static function that attaches an object to a surface, setting its position and parent and applying rotation so it appears to be appropriately stuck onto the surface. Since this function has 3 parameters (the object’s transform, a RaycastHit for the surface and a Vector3 for the direction), it cannot be activated from a UnityEvent assigned in the Inspector, so I added an extra non-static function where the developer only needs to specify the projectile, and the function will use the projectile’s impact information and a default direction of forward, to automatically stick it and rotate its forward axis outwards based on the surface’s normal.
I also added two extra non-static functions, one that simply spawns an object at the point without rotation, and one that sticks the projectile itself into the target (for cases where the projectile itself needs to continue existing and performing functions once hitting a surface).
I then implemented a usage of this function by adding an event using it to the bullet prefab, so upon impact it spawns shot debris particles that fly off the surface of the impact point.