Experimented with a script for enemy AI
The script does not feature much, presently – it includes a reference to a Character script and a Health script, functions for looking around and an unfinished property for obtaining objects within the AI’s field of view. I realised that I have other, more important things I need to work on before reaching this level of detail and complexity.
I also made a simple placeholder enemy prefab to apply these scripts to.