Fixed a bug preventing projectiles from aiming properly if the raycast check does not land
The position for specifying where the bullet should be aimed if the raycast does not land (projectileDirection) was simply a direction value. Since the origin of the shot was not added, the game interpreted it as the direction of the shot from 0,0,0 regardless of the player's position. I fixed this by adding the player's aim origin to the value.

I also added several comments for improved clarity.