Created AIGridPoints class to improve AI pathfinding
The purpose of this class is to provide AI entities with a handy resource of positions to sample for pathfinding, pre-calculated to improve performance.
This class is a MonoBehaviour that exists in the scene as a singleton. During Awake(), it uses a Bounds value to sweep the level with a grid, using scanning a massive array of evenly spaced points on the level geometry using Physics.RaycastAll (if I used regular raycasts it would only penetrate the topmost floor and be inadequate for multi-levelled maps) and save them in a Vector3 list.
The class also includes two functions for easily obtaining a set of points under certain criteria.
The first function, GetPoints(), uses an origin position, and a minimum and maximum radius as parameters. The function will handily copy the list of grid points and erase all entries that are too close or too far away, for a very accurate cluster of points.
The second function, GetSpecificNumberOfPoints, also includes a quantity integer as a parameter in case the code only wants to sample a small number of positions to prevent poor performance. After gaining a list of points from the first function, it runs a for loop based on the quantity integer to obtain that particular amount of values. Instead of using the for loop index to source the values, each iteration of the loop divides the (larger) length of the original list by the (smaller) quantity integer (if the quantity is larger than the list, it just returns the whole list), multiplies by the index of the for loop and rounds to obtain an index for the value. This allows the loop to sample the correct number of points from all across the list, resulting in a very uniform distribution of points.
I may not be at the stage of completing fully functioning AI yet, but this class is an important element I needed to include, and I'm quite pleased with how it turned out.