Created complex line of sight check function
This static function acts as a more complex pathfinding check, that creates a sweeping grid of raycasts to check if a collider is partially concealed.
The function creates up and right directions, perpendicular to the direction the collider is in from the origin. It then multiplies these direction values by the collider’s bounds’ extents’ magnitude, creates inverted versions and calculates the distances to create a zone to blanket with raycasts. It does more math to calculate a grid of positions, using a float for the specific spacing of raycasts, and then performs a sweep of raycasts to check if one of the hits is the collider being checked for. This method will probably not as performant as would be ideal for such a game, but after testing I can conclude that it does its job perfectly.
To test the function, I also created a FieldOfView class. It is currently just used to periodically cast the aforementioned line of sight check between itself and a collider in the scene, but I will eventually use it to hold an array of field-of-view related stats and functions.