Implemented cosmetic drag, tilt and sway for movement controller
The player’s movement controller class now has some extra animations when moving – the player’s weapons will drag and tilt in certain directions when moving, and lag slightly behind when they look around, to make them seem more real and physical. No actual physics are involved, however.
The code works by moving and rotating a single transform that is a child of the player’s upper body transform, which is useful as it allows everything the player is holding to be affected by these values. During LateUpdate(), the code will obtain the amount that the player has moved and rotated in the last frame, multiply the magnitudes by different max values for the three different types of cosmetic rotation, and clamp them between zero and one. Those values are then used to lerp position and rotation values between zero and the direction or rotation values multiplied by maximum possible angle values, and added to two Vector3 values – one for local position and the other for local rotation. These two values are then applied to the animation transform using SmoothDamp functions.