Implemented static EventHandler class to allow easy transmission of information about in-game events.
If an object in the game wishes to act upon certain common actions being performed, such as player interactions, characters taking damage or dying, etc., they can take the function they want to run and add it to/remove it from an appropriate delegate in the EventHandler using EventHandler.Subscribe(). When one of those actions is performed, it will transmit a class containing the necessary information using EventHandler.Transmit(), which will then run all functions in the appropriate delegate. Each action has a different message class, and a different delegate in the EventHandler.