Enemy health meter in HUD
When the player has their reticle trained on an entity that can be damaged, a meter will display over their heads to communicate their current health status.
If the player’s reticle is currently aiming at an enemy hitbox, code will run to compile the bounds of the enemy’s hitboxes to make a single combined Bounds struct. A world position is produced by obtaining the bounds’ centre, and moving outwards for a distance of the extents’ magnitude (so the point is always outside the entity’s bounds and doesn’t cover up any part of the enemy), in the direction of the camera’s up vector (so no matter how the camera is rotated the point will always be above the enemy in the player’s view).
This world position is converted to a screen position using Camera.WorldToScreenPoint(), then into an anchored position on the canvas using a proprietary function named MiscFunctions.ScreenToAnchoredPosition().
A health meter object attached to the heads-up display canvas is enabled or disabled depending on if an enemy is sighted, and then set to this anchored position and updated to reflect the enemy’s health values.
This system allows for a simple and intuitive way for the player to detect the remaining health of an enemy.
Additionally, in order to make this work, I slightly updated the object detection code in the player’s HUD, to specify a variable detection range. I implemented this change in case the player is using a weapon with a very different range to the default observation range.