TimedCosmeticEffect and LerpingCosmeticEffect system for code-based aesthetics
Created a system for easily handing cosmetic effects that change via code-based means, such as changing colours or scale. The system consists of two classes. The first, TimedCosmeticEffect, contains public functions for playing, pausing, resuming and stopping the effects, with a duration variable and a UnityEvent with a single float parameter. While playing, the effect will make a timer float count up from zero to one, and use it to invoke the UnityEvent. The float can also be altered directly, for visual effects that don’t simply occur over time but change back and forth based on another gameplay factor.
The UnityEvent is then used to invoke LerpingCosmeticEffect, an abstract MonoBehaviour class containing an AnimationCurve and an abstract function for altering values using a float as a parameter. There are several variants of this class for altering different effects (currently there are four, for the three different transform values plus image colour). These can be assigned values and slotted into a TimedCosmeticEffect component’s UnityEvent, and when that component’s state is changed each effect will change accordingly. While I am sure there are ways to make this system better, I am pleased with the way it performs.