Quickly overhauled weapon animation system to make it usable for things other than weapon animations
I took the code present in the Weapon class, and moved it into a new MonoBehaviour named SimpleLerpAnimator, adding a reference to the transform being animated. I also took the variables in SimpleWeaponAnimation and put them into a struct named SimpleLerpAnimation, adding an array of them to the SimpleLerpAnimator class. Finally, I added a function in the SimpleLerpAnimator that will easily play a desired animation from the array, using an index value. I also removed the references to the weapon model and model transform in the Weapon class, as they are now redundant.
The effect now works much better, with developers being able to add any number of animations via the array, without having to make separate Monobehaviours for every possible animation, and with the ability to use said animations on anything, not just weapons.