Created target prefab for testing
This object resembles a human-shaped firing range target and exists as a demo object to show off the damage physics. It has a health meter and different hitboxes, and can be shot by the player, and when it runs out of health it will 'die', playing a short falling over animation before rising back up with its health resurrected.