ADS functionality implemented
The guns in my FPS now have proper ADS functionality. I had to change a few of the mechanics compared to how I did it in my previous project, but my new version is far better.
For ADS functions where the player actually looks through a scope, I have code and variables to show the correct view in a much more realistic way. The scope object has a camera and a world-space canvas with an Image object attached. During Awake(), if these variables are assigned, a RenderTexture is generated and assigned to the Image object, and the camera is set to output to the RenderTexture. The viewing camera’s FOV is set to the player’s FOV divided by the magnification value, to be properly zoomed in relative to the player’s field of view. This way, when the player aims down the scope they will get a properly zoomed in view more akin to peering through a real gun scope.
If the viewing camera and sight picture are not assigned (determined by a property bool named IsScope for easy referencing in multiple use cases), the player’s regular camera FOV will be zoomed in instead using a lerping function, for guns that do not have a proper scope but where I want to implement zooming anyway. This is not strictly a realistic mechanic, but I feel it enhances gameplay enough to offset it for certain kinds of weapons. I choose to interpret it as a representation of the player-character being a trained soldier, who can peer through unmagnified ironsights far better than an ordinary person would.
Changed the name of the script from GunOptics to GunADS, as it is now used for all kinds of ADS functionality, not just for weapons with scopes.
Instead of using the system I currently have for other animations, the animation involves a separate lerping mechanic in the ADS script’s LateUpdate() function, that runs if IsAiming or IsTransitioning is set to true. The code does not need to run if the player is not aiming, and is not transitioning back to the hipfiring position. When lerping the animations, instead of specifying a regular ADS position, the gun ADS script has a transform variable for a relative head orientation. The gun will then move so this transform has the exact same position and rotation as the player. This exists to ensure that the reticle or ironsights on the weapon properly lines up with the centre of the player’s vision and accurately represents where they are aiming.
The player’s normal reticle will easily disable when they are aiming down the weapon’s sights, or if a variable on the scope is enabled to hide it anyway. This variable exists in case I want a weapon that the player is not supposed to be able to hipfire with, such as a slow sniper rifle.
I added a circle graphic for masking the scope sight picture, and a prettier-looking chevron reticle for both basic viewing and for the scope graphic. I believe I will eventually make some new reticles that the player can choose, but I might instead have it change based on the current weapon mode the player is using.
I made the scope on the player’s rifle into a Placeholder Scope prefab, so it can be easily added to other weapons.
This system has some flaws that need to be accounted for. When aiming down the sights the player’s gun does not move in any way and it looks stiff and awkward. Additionally, when I use a mask to ensure the sight picture is a shape that matches the actual scope model without clipping outwards, it causes every image that has no texture assigned to duplicate the scope view. However, those are not gamebreaking issues, and I am very pleased with how well the system currently works.