Field of view now updates properly to match the value specified in the MovementController class.
This functionality is necessary because the player camera's FOV can be changed by external factors such as using ADS on weapons, so there needs to be a default value specified elsewhere. Since this is the value being used, the player's FOV needs to update at the start to reflect that.