Changed mechanic for setting TimedCosmeticEffect time from a property to a function
As a property, its true purpose was somewhat ambiguous, as it exists simply as a means to update the values directly via other sources.
I also added an OnValidate() function with code to update the values of its effects in the editor, to ensure that the function resembles what it should at the start of the animation.