Muzzle flash effects on weapons
The player’s guns now have a simple muzzle flash effect that plays when the weapon is fired. This uses my TimedCosmeticEffect system to rapidly expand and shrink a simple box mesh with an attached material to resemble a muzzle flash. The effect’s size changes so quickly that it is not easy to notice that the shape is just a box, and the muzzle flash will move with the gun, two side effects that are not realistic but easily ignored. The effect also uses a new effect, LightDataEffect, that warps the size of an attached light component to dramatically illuminate the area when the gun is fired.