Started creating MonoBehaviour scripts for menus
The MenuWindow script attaches to a Canvas and represents a single menu screen that can be navigated to. It contains references to its children, its immediate parent and its root for hierarchy purposes, as well as miscellaneous functions activated from a menu, such as loading scenes and quitting the game. This class will also include better code for switching between screens in a hierarchy using canvas groups, but that has not been implemented yet.

The MenuInteractable class serves to make it easier to create menu buttons. If this script is present, I can easily change the name of an interactable object in the Inspector and have its title text change, making organisation easy for both the player and the developer. It also includes info for descriptive graphics and flavour text, that can be referenced to easily show tooltips for menu options.