Fixed quaternion errors with player cosmetic animations
The desired cosmetic positions and rotations were being turned into NaN values and causing errors, when the player paused and then resumed the game. I traced this back to code that divides a number by Time.deltaTime. If the game is paused, this will result in the code attempting to divide by zero, causing issues. I fixed this by adding a simple check to return the cosmetic functions early if Time.deltaTime is less than or equal to zero.