Added bobbing animations to player walk cycle
In the walk cycle animation code, as well as an if statement that checks the step cycle timer to play step noises, I added an extra bit of code that uses two animation curves to produce a smoothly interpolating Vector3 position based on how far the player is through their walk cycle. This position value is then added to the torsoPosition value, in the same way as the TorsoDrag function, to create a simple but very nice looking bobbing animation.
I also slightly tweaked the way the walk cycle time is calculated. Instead of just having a fixed duration float, the length is calculated by the player’s stride length, multiplied by the number of steps in the cycle and then divided by the player’s current movement speed. This means the walk animation will properly adjust based on the player’s movement speed, and steps will stay consistent relative to the environment the player is walking in.