Tweaked movement controller code so walk cycle animations play through properly if player only moves a small distance
If the player starts moving but stops shortly before the full walk cycle is completed, the walk cycle will now continue until it is finished, so players will not mysteriously make no noise if they only move in small increments. I made this change by having the if statement that activates the walk cycle check if, if the player is not inputting movement, if the walk cycle timer does not equal zero. This ensures the function will continue running until the current cycle is played out and the timer returns to zero.