Added Play On Awake option to TimedCosmeticEffect, and made fix to ensure animations do not start partially when disabled.
This bool will allow the effect to play immediately when the object is loaded, without needing to be manually triggered. Good for ambient background effects.
Additionally, I added a check in LateUpdate() to end the function prematurely if the script is disabled. This is necessary because if the script is disabled, LateUpdate() will still run and change the animations because it was enabled at the start of the frame. This caused problems with very quick animations that are supposed to not play automatically, which would still play for a frame and be visibly stuck in time.