Added switching functionality to menus
During Start(), each MenuWindow code will find its parents and children in the hierarchy. If a MenuWindow does not have a parent, it is the root, and will automatically assign itself as such to itself and all other menus in the hierarchy.
When I want to implement functionality to switch to a different menu window, I can add a SwitchWindow() function as a listener, and assign a MenuWindow. The code will then cleanly disable all windows except for the desired window and its parents, then alter the CanvasGroup alpha of each parent to hide them from being viewed and interacted with. I had to use the CanvasGroup for this because simply disabling a canvas or its gameobject will disable all child canvas objects, meaning if the current window has a parent I cannot disable it without disabling the window I actually want to look at.
I also implemented two extra functions – ReturnToParentMenu() and ReturnToRoot(), which will automatically switch to the parent or root menu respectively without any parameters to specify. A reference for a ‘back’ button is present in the MenuWindow script, which has ReturnToParentMenu() assigned as a listener in Awake(). These functions are useful because it allows me to drop a menu prefab into a hierarchy, and it will automatically have functionality to return to its parent without me needing to assign listeners to menu buttons.