Created a main menu scene
This scene currently includes a title splash and a main menu featuring two options - one to go to the testing target range scene, and one to quit the game. There are also options for starting a new game, loading an existing one, viewing achievements and updating options, but none of those work yet.